Reordered a bit codebase and cleaned up event categories retrieval
